+++ src/ExternalLink.tsx
@@ -7,12 +7,13 @@ import Icon from './Icon';
 
 export type ExternalLinkProps = {
   children: React.ReactNode;
+  className?: string;
   to: string;
 };
 
-function ExternalLink({ children, to }: ExternalLinkProps) {
+function ExternalLink({ children, className = undefined, to }: ExternalLinkProps) {
   return (
-    <a href={to} rel="noopener noreferrer" target="_blank">
+    <a className={className} href={to} rel="noopener noreferrer" target="_blank">
       {children} <Icon icon="open_in_new" size={18} />
     </a>
   );